+++ vol_test.c
@@ -156,11 +156,11 @@ int main(int argc, char **argv)
     vol_test_run();
 
     HDprintf("%ld/%ld (%.2f%%) VOL tests passed with VOL connector '%s'\n",
-            (long) n_tests_passed_g, (long) n_tests_run_g, ((float) n_tests_passed_g / n_tests_run_g * 100.0), vol_connector_name);
+            (long) n_tests_passed_g, (long) n_tests_run_g, ((float) n_tests_passed_g / (float) n_tests_run_g * 100.0), vol_connector_name);
     HDprintf("%ld/%ld (%.2f%%) VOL tests did not pass with VOL connector '%s'\n",
-            (long) n_tests_failed_g, (long) n_tests_run_g, ((float) n_tests_failed_g / n_tests_run_g * 100.0), vol_connector_name);
+            (long) n_tests_failed_g, (long) n_tests_run_g, ((float) n_tests_failed_g / (float) n_tests_run_g * 100.0), vol_connector_name);
     HDprintf("%ld/%ld (%.2f%%) VOL tests were skipped with VOL connector '%s'\n",
-            (long) n_tests_skipped_g, (long) n_tests_run_g, ((float) n_tests_skipped_g / n_tests_run_g * 100.0), vol_connector_name);
+            (long) n_tests_skipped_g, (long) n_tests_run_g, ((float) n_tests_skipped_g / (float) n_tests_run_g * 100.0), vol_connector_name);
 
 done:
     H5close();
